# docs/openapi.yaml
openapi: 3.1.0
info:
  title: PROOF API
  description: PROOF 서비스 REST API 명세서
  version: 1.0.0
  contact:
    name: PROOF Team
servers:
  - url: http://localhost:8000
    description: 로컬 개발 서버
...
paths:
...
  /api/job-simulation/start:
    post:
      tags:
        - Job Simulation
      summary: 직무 시뮬레이션 시작
      description: 직무 시뮬레이션을 시작합니다
      parameters:
        - $ref: '#/components/parameters/UserIdHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - department
              properties:
                department:
                  type: string
                  enum: [Business, Economics, Statistics]
                  description: 학과
                  example: "Business"
      responses:
        '200':
          description: 시뮬레이션 시작
          content:
            application/json:
              schema:
                type: object
                properties:
                  session_id:
                    type: string
                    format: uuid
                    example: "123e4567-e89b-12d3-a456-426614174000"
                  department:
                    type: string
                    example: "Business"
                  total_questions:
                    type: integer
                    example: 15
  /api/job-simulation/submit:
    post:
      tags:
        - Job Simulation
      summary: 시뮬레이션 답변 제출
      description: 시뮬레이션 결과를 제출하고 분석 결과를 받습니다
      parameters:
        - $ref: '#/components/parameters/UserIdHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - session_id
                - answers
              properties:
                session_id:
                  type: string
                  format: uuid
                answers:
                  type: array
                  items:
                    type: object
                    properties:
                      question_id:
                        type: integer
                      selected_option:
                        type: integer
      responses:
        '200':
          description: 결과 분석 완료
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobSimulationResult'
  /api/job-simulation/results:
    get:
      tags:
        - Job Simulation
      summary: 시뮬레이션 결과 목록
      description: 사용자의 시뮬레이션 결과 목록을 조회합니다
      parameters:
        - $ref: '#/components/parameters/UserIdHeader'
      responses:
        '200':
          description: 결과 목록
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/JobSimulationResult'
  /api/job-simulation/results/{result_id}:
    get:
      tags:
        - Job Simulation
      summary: 시뮬레이션 결과 상세
      description: 특정 시뮬레이션 결과를 조회합니다
      parameters:
        - $ref: '#/components/parameters/UserIdHeader'
        - name: result_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 결과 상세
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobSimulationResult'

  # ==================== Career Survey ====================
...
components:
  parameters:
    UserIdHeader:
      name: x-user-id
      in: header
      required: true
      description: 사용자 ID
      schema:
        type: string
        example: "john_doe"
  schemas:
...
    JobSimulationResult:
      type: object
      properties:
        result_id:
          type: string
          format: uuid
        user_id:
          type: string
        department:
          type: string
          enum: [Business, Economics, Statistics]
        scores:
          type: object
          description: 직무별 점수
          example:
            MKT: 85
            PM: 75
            DATA: 65
            DEV: 45
            DESIGN: 55
            HR: 70
            FIN: 60
        top_job:
          type: string
          example: "마케터"
        strengths:
          type: array
          items:
            type: string
          example: ["시장 분석 능력", "커뮤니케이션", "창의적 사고"]
        recommendations:
          type: array
          items:
            type: string
          example: ["마케팅 공모전 참여", "브랜드 분석 프로젝트"]
        ai_analysis:
          type: string
          example: "마케팅 분야에서 강점을 보이고 있습니다..."
        completed_at:
          type: string
          format: date-time
...